+++ extra/gen_rf_survey.py
@@ -3,12 +3,8 @@
 from pathlib import Path
 
 from shepherd_core import WebClient
+from shepherd_core import data_models as sm
 from shepherd_core import logger
-from shepherd_core.data_models import GpioTracing
-from shepherd_core.data_models.content import EnergyEnvironment
-from shepherd_core.data_models.content import Firmware
-from shepherd_core.data_models.experiment import Experiment
-from shepherd_core.data_models.experiment import TargetConfig
 from shepherd_core.data_models.task import TestbedTasks
 
 if __name__ == "__main__":
@@ -31,29 +27,25 @@ if __name__ == "__main__":
         path_task.mkdir(parents=True)
 
     # RF-Survey
-    xp1 = Experiment(
+    xp1 = sm.Experiment(
         name="rf_survey",
         comment="generate link-matrix",
         duration=8 * 60,
         target_configs=[
-            TargetConfig(
+            sm.TargetConfig(
                 target_IDs=list(range(1, 12)),
                 custom_IDs=list(range(1, 32)),
                 # ⤷ note: traffic bench expects node 1 as root-node
-                energy_env=EnergyEnvironment(name="eenv_static_3000mV_50mA_3600s"),
-                firmware1=Firmware.from_firmware(
+                energy_env=sm.EnergyEnvironment(name="eenv_static_3000mV_50mA_3600s"),
+                firmware1=sm.Firmware.from_firmware(
                     file=path_fw / "nrf52_rf_survey/build.elf",
                     embed=False,
                     owner="Ingmar",
                     group="NES_Lab",
                 ),
-                firmware2=Firmware(name="msp430_deep_sleep"),
+                firmware2=sm.Firmware(name="msp430_deep_sleep"),
                 power_tracing=None,
-                gpio_tracing=GpioTracing(
-                    duration=0,  # this affects only the gpio-tracing itself
-                    uart_decode=True,  # TODO: enables logging uart from userspace, temp solution
-                    uart_baudrate=115_200,
-                ),
+                uart_tracing=sm.UartTracing(baudrate=115_200),
             )
         ],
     )