+++ shepherd_core/shepherd_core/data_models/content/virtual_harvester.py
@@ -23,9 +23,9 @@ from .energy_environment import EnergyDType
 class AlgorithmDType(str, Enum):
     """Options for choosing a harvesting algorithm."""
 
-    direct = disable = neutral = "neutral"
-    isc_voc = "isc_voc"
-    ivcurve = ivcurves = ("ivcurve",)
+    direct = disable = neutral = "neutral"  # for just using IVTrace / samples
+    isc_voc = "isc_voc"  # only recordable ATM
+    ivcurve = ivcurves = ivsurface = "ivcurve"
     constant = cv = "cv"
     # ci .. constant current -> is this desired?
     mppt_voc = "mppt_voc"
@@ -37,7 +37,7 @@ class VirtualHarvesterConfig(ContentModel, title="Config for the Harvester"):
     """A vHrv makes a source-characterization (i.e. ivcurve) usable for the vSrc.
 
     Mostly used when the file-based energy environment of the virtual source
-    is not already supplied as pre-harvested ivsample-stream.
+    is not already supplied as pre-harvested ivtrace.
     """
 
     # General Metadata & Ownership -> ContentModel