+++ shepherd_core/shepherd_core/fw_tools/converter.py
@@ -28,7 +28,8 @@ def firmware_to_hex(file_path: Path) -> Path:
         return elf_to_hex(file_path)
     if is_hex(file_path):
         return file_path
-    raise FileNotFoundError("FW2Hex: unknown file '%s', it should be ELF or HEX", file_path.name)
+    msg = (f"FW2Hex: unknown file '{file_path.name}', it should be ELF or HEX",)
+    raise FileNotFoundError(msg)
 
 
 @validate_call
@@ -98,10 +99,12 @@ def extract_firmware(data: Union[str, Path], data_type: FirmwareDType, file_path
         elif data_type == FirmwareDType.path_hex:
             file = file_path.with_suffix(".hex")
         else:
-            raise ValueError("FW-Extraction failed due to unknown datatype '%s'", data_type)
+            msg = "FW-Extraction failed due to unknown datatype '{data_type}'"
+            raise ValueError(msg)
         if not file.parent.exists():
             file.parent.mkdir(parents=True)
         shutil.copy(data, file)
     else:
-        raise ValueError("FW-Extraction failed due to unknown data-type '%s'", type(data))
+        msg = f"FW-Extraction failed due to unknown data-type '{type(data)}'"
+        raise ValueError(msg)
     return file